requirements.txt
numpy
scipy
matplotlib
pandas
pyserial
openpyxl
control
Pillow
